mcp_cli_auth_tool/requirements.txt
# Base requirements
fastapi
uvicorn
requests
google-auth-oauthlib
python-dotenv
pyjwt
cryptography
